Added Title and edit options
I am goign to start working on this again with me starting off by adding  a title element which adds text that is inside a h1 tag. There is also now a dropdown for the text editing. I will also be refining how you add new page elements to display.